map re-renders after location is selected in dropdown
