+++ frontend/src/pages/HomePage.jsx
@@ -1,5 +1,7 @@
 import Feed from "../components/Feed";
 import Map from "../components/Map";
+import { useState } from "react";
+import { neighborhoodCenters, boroughCenters } from "../utils/neighborhoodCenters";
 
 /**
  * HomePage should be shown on '/'
@@ -9,10 +11,40 @@ import Map from "../components/Map";
  */
 
 export default function HomePage() {
+  const [filterType, setFilterType] = useState("all");
+  const [filterValue, setFilterValue] = useState("");
+  const [mapCenter, setMapCenter] = useState({ lat: 40.65798, lng: -74.005439 });
+  const [mapZoom, setMapZoom] = useState(12);
+
+  const handleFilterChange = (e) => {
+    const value = e.target.value;
+    if (value.includes(":")) {
+      const [type, val] = value.split(":");
+      setFilterType(type);
+      setFilterValue(val);
+      if (type === "borough" && boroughCenters[val]) {
+        setMapCenter(boroughCenters[val]);
+        setMapZoom(12);
+      } else if (type === "neighborhood" && neighborhoodCenters[val]) {
+        setMapCenter(neighborhoodCenters[val]);
+        setMapZoom(15);
+      }
+    } else {
+      setFilterType("all");
+      setFilterValue("");
+      // Optionally reset map center/zoom here
+    }
+  };
+
+  const handleMapMove = (location, zoom = 17) => {
+    setMapCenter(location);
+    setMapZoom(zoom);
+  };
+
   return (
     <div className="homepage">
-      <Feed />
-      <Map />
+      <Feed filterType={filterType} filterValue={filterValue} onFilterChange={handleFilterChange} onMapMove={handleMapMove} />
+      <Map mapCenter={mapCenter} mapZoom={mapZoom} onMapMove={handleMapMove} />
     </div>
   );
 }
